+++ enigma-test-spring-boot-starter/src/main/java/com/rdlts/enigma/test/domain/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * 实用测试对象
+ */
+package com.rdlts.enigma.test.domain;