+++ enigma-test-spring-boot-starter/src/test/java/com/rdlts/enigma/test/bdd/demo/EnigmaBracerQuestBddContext.java
@@ -1,0 +1,15 @@
+package com.rdlts.enigma.test.bdd.demo;
+
+import com.rdlts.enigma.test.core.EnigmaCucumberSpringBootBasedTest;
+
+/**
+ * EnigmaBracerQuestBddContext
+ * 游击士支援任务评级 BDD 示例的 Spring 上下文桥接类。
+ *
+ * @author wangjialong
+ * @since 2026/05/11 16:05
+ */
+public class EnigmaBracerQuestBddContext extends EnigmaCucumberSpringBootBasedTest {
+}
+
+