+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### 新增
 - 版本升级至 1.2.0-SNAPSHOT
+- enigma-test-spring-boot-starter 新增 BDD 测试组件与游击士支援任务评级 demo
 
 ## 1.1.1-RELEASE
 - 增加 maven版本更新脚本